♻️ unreadCount를 부모에서 계산해서 props로 전달하도록 변경
